Separated ResultPage into multiple compoments and used react-bootstrap grid system to adjust layout
