+++ react-app/src/components/ResultHeader.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function ResultHeader() {
+  return (
+    <div>
+      <h1 className="page-title">Results Header Will Go Here</h1>
+    </div>
+  );
+}
+
+export default ResultHeader;
